tests/GPTで簡易的なSlackチャットBotの開発/requirements.txt
python-dotenv
slack-sdk
transformers